# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RepoPilot Frontend–Backend API",
    "description": "Contract of HTTP APIs that the frontend (`frontend/src/lib/api.ts`, `auth.ts`) calls against the backend. Derived from frontend clients and response TypeScript types.",
    "version": "4.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local backend (dev). Vite proxies `/api` and `/auth` from :5173 to this origin."
    }
  ],
...
  "paths": {
...
    "/api/user/setting/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get email notification preferences",
        "operationId": "fetchNotificationSettings",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Notification preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettings"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "Notifications"
        ],
        "summary": "Update email notification preferences",
        "operationId": "updateNotificationSettings",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSettingsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettings"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/user/setting/notifications/test": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Send a stub test notification email (no real delivery)",
        "operationId": "sendTestNotification",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Stub send result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationTestResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "GitHub OAuth access token",
        "description": "Token from GitHub OAuth (`Authorization: Bearer <token>`). Frontend stores it in localStorage and sends it via `apiFetch` when present."
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid Bearer token. Frontend clears auth and redirects to `/login`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "detail": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": true
                }
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          }
        }
      },
...
      "NotificationSettings": {
        "type": "object",
        "required": [
          "email",
          "enabled",
          "notifyOnKnowledgeBuild",
          "notifyOnIssueAnalysis",
          "notifyOnWikiReady",
          "deliveryMode",
          "updatedAt"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "notifyOnKnowledgeBuild": {
            "type": "boolean"
          },
          "notifyOnIssueAnalysis": {
            "type": "boolean"
          },
          "notifyOnWikiReady": {
            "type": "boolean"
          },
          "deliveryMode": {
            "type": "string",
            "enum": [
              "stub"
            ],
            "description": "Only stub delivery is implemented; no real SMTP send"
          },
          "updatedAt": {
            "type": "string"
          },
          "lastTestAt": {
            "type": "string"
          },
          "lastTestMessage": {
            "type": "string"
          }
        }
      },
      "NotificationSettingsUpdate": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "notifyOnKnowledgeBuild": {
            "type": "boolean"
          },
          "notifyOnIssueAnalysis": {
            "type": "boolean"
          },
          "notifyOnWikiReady": {
            "type": "boolean"
          }
        }
      },
      "NotificationTestResult": {
        "type": "object",
        "required": [
          "success",
          "deliveryMode",
          "message",
          "sentAt"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "deliveryMode": {
            "type": "string",
            "enum": [
              "stub"
            ]
          },
          "message": {
            "type": "string"
          },
          "sentAt": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
...
    }
  }
}